requirements.txt
python-telegram-bot==20.6
python-dotenv
openai
aiosqlite
